Reposition elements from webpage cover
