+++ src/styles/components/Block.pcss
@@ -76,8 +76,6 @@
   }
 
   &.first {
-    margin-top: 0;
-
     & .head {
       padding: 0;
       border-bottom: none;
