+++ src/styles/components/Topline.pcss
@@ -1,4 +1,5 @@
 .Topline {
+  margin-bottom: 10px;
 
   & hr {
     display: none;
