--- src/styles/components/WebPageCover.pcss
@@ -1,22 +1,0 @@
-.WebPageCover {
-  background-color: var(--grey-lightest);
-  padding: var(--webpage-padding);
-  position: relative;
-  height: 280px;
-
-  & .ResourceImage {
-    width: 200px;
-    height: 200px;
-    max-height: 200px;
-    transition: opacity 0.5s;
-
-    & img {
-      padding: 10px;
-    }
-  }
-
-  & .webPageCoverImage .missingImg {
-    background: var(--white);
-    color: var(--grey-light);
-  }
-}
